Renamed gql to have .ts extension, moved to lib.
