+++ src/schema/__tests__/collection.test.js
@@ -1,7 +1,7 @@
 import { resolve } from "path"
 import { readFileSync } from "fs"
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 import { isEqual } from "lodash"
 
 const gravityData = {
