+++ src/schema/__tests__/ecommerce/create_order_with_artwork_mutation.test.js
@@ -3,7 +3,7 @@ import { runQuery } from "test/utils"
 import { mockxchange } from "test/fixtures/exchange/mockxchange"
 import sampleOrder from "test/fixtures/results/sample_order"
 import exchangeOrderJSON from "test/fixtures/exchange/order.json"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 let rootValue
