+++ src/schema/artists/__tests__/popular.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable promise/always-return */
-import gql from "test/gql"
+import gql from "lib/gql"
 import { runQuery } from "test/utils"
 
 describe("Popular Artists", () => {
