+++ src/schema/me/__tests__/artwork_inquiries.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable promise/always-return */
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 describe("Me", () => {
   describe("ArtworkInquiries", () => {
