+++ src/schema/me/__tests__/consignments/add_asset_to_submission_mutation.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable promise/always-return */
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 describe("addAssetToConsignmentSubmission", () => {
   it("creates a submission and returns its new data payload", () => {
