+++ src/schema/me/__tests__/consignments/submissions.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable promise/always-return */
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 describe("submissions", () => {
   it("asks for a user's submissions", () => {
