+++ src/schema/me/__tests__/consignments/update_submission_mutation.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable promise/always-return */
 import { runAuthenticatedQuery } from "test/utils"
 import { config as updateSubmissionMutation } from "schema/me/consignments/update_submission_mutation"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 describe("UpdateSubmissionMutation", () => {
   it("includes the id param", () => {
