+++ src/schema/me/__tests__/followed_artists.test.js
@@ -2,7 +2,7 @@
 import { resolve } from "path"
 import { readFileSync } from "fs"
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 it("returns artists for a user", () => {
   const artworksPath = resolve(
