+++ src/schema/me/__tests__/save_artwork.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable promise/always-return */
 import { runAuthenticatedQuery } from "test/utils"
-import gql from "test/gql"
+import gql from "lib/gql"
 
 describe("SaveArtworkMutation", () => {
   it("saves an artwork", () => {
